requirements.txt
Django==4.2.7
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-cleanup==8.0.0
django-ckeditor==6.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
whitenoise==6.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9  # For PostgreSQL
dj-database-url==2.1.0
django-storages==1.14.2
boto3==1.28.38  # For AWS S3 storage
django-allauth==0.57.0  # For social authentication
